Extract CompleteContext from CheckoutContext
