+++ src/Sylius/Behat/Resources/config/suites/ui/account/customer.yml
@@ -35,6 +35,7 @@ default:
                 - sylius.behat.context.ui.shop.account
                 - sylius.behat.context.ui.checkout
                 - sylius.behat.context.ui.checkout.addressing
+                - sylius.behat.context.ui.checkout.complete
                 - sylius.behat.context.ui.shop.cart
                 - sylius.behat.context.ui.shop.currency
             filters:
