+++ src/Sylius/Behat/Resources/config/suites/ui/checkout/paying_for_order.yml
@@ -28,6 +28,7 @@ default:
 
                 - sylius.behat.context.ui.checkout
                 - sylius.behat.context.ui.checkout.addressing
+                - sylius.behat.context.ui.checkout.complete
                 - sylius.behat.context.ui.checkout.order_details
                 - sylius.behat.context.ui.checkout.payment
                 - sylius.behat.context.ui.checkout.shipping
