+++ src/Sylius/Behat/Resources/config/suites/ui/order/managing_orders.yml
@@ -49,6 +49,7 @@ default:
                 - sylius.behat.context.ui.shop.currency
                 - sylius.behat.context.ui.checkout
                 - sylius.behat.context.ui.checkout.addressing
+                - sylius.behat.context.ui.checkout.complete
                 - sylius.behat.context.ui.channel
                 - sylius.behat.context.ui.shop.cart
             filters:
